+++ app/app.go
@@ -36,7 +36,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/bank"
 
 	"github.com/notional-labs/composable/v6/app/keepers"
-	"github.com/notional-labs/composable/v6/app/upgrades/v6_5_3"
+	"github.com/notional-labs/composable/v6/app/upgrades/v6_5_4"
 
 	// bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
 
@@ -147,7 +147,7 @@ var (
 	// https://github.com/CosmWasm/wasmd/blob/02a54d33ff2c064f3539ae12d75d027d9c665f05/x/wasm/internal/types/proposal.go#L28-L34
 	EnableSpecificProposals = ""
 
-	Upgrades = []upgrades.Upgrade{v6_5_3.Upgrade}
+	Upgrades = []upgrades.Upgrade{v6_5_4.Upgrade}
 	Forks    = []upgrades.Fork{}
 )
