+++ app/upgrades/v6_5_4/constants.go
@@ -1,4 +1,4 @@
-package v6_5_3
+package v6_5_4
 
 import (
 	store "github.com/cosmos/cosmos-sdk/store/types"
@@ -7,7 +7,7 @@ import (
 
 const (
 	// UpgradeName defines the on-chain upgrade name for the composable upgrade.
-	UpgradeName = "v6_5_3"
+	UpgradeName = "v6_5_4"
 )
 
 var Upgrade = upgrades.Upgrade{
